+++ src/client/app/heroes/heroes.module.ts
@@ -5,12 +5,12 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { HeroesRoutingModule } from './heroes-routing.module';
 import { HeroDetailComponent } from './hero-detail/hero-detail.component';
 import { HeroesComponent } from './heroes/heroes.component';
+import { HeroesV1Component } from './heroes/heroes.v1.component';
 import { SharedModule } from '../shared/shared.module';
 import { HeroListComponent } from './hero-list/hero-list.component';
 
 @NgModule({
   imports: [CommonModule, ReactiveFormsModule, HeroesRoutingModule, SharedModule],
-  exports: [HeroesComponent, HeroDetailComponent],
-  declarations: [HeroesComponent, HeroDetailComponent, HeroListComponent]
+  declarations: [HeroesComponent, HeroesV1Component, HeroDetailComponent, HeroListComponent]
 })
 export class HeroesModule {}